requirements.txt
discord
Pillow
mysql-connector-python
redis